Added example for auth middleware
